+++ src/MusicalScore/Graphical/VexFlow/VexFlowVoiceEntry.ts
@@ -9,6 +9,7 @@ import { Note } from "../../VoiceData/Note";
 import { ColoringModes } from "../../../Common/Enums/ColoringModes";
 import { GraphicalNote } from "../GraphicalNote";
 import { EngravingRules } from "../EngravingRules";
+import { NoteHeadShape } from "../../VoiceData/Notehead";
 
 export class VexFlowVoiceEntry extends GraphicalVoiceEntry {
     private mVexFlowStaveNote: VF.StemmableNote;
@@ -86,8 +87,8 @@ export class VexFlowVoiceEntry extends GraphicalVoiceEntry {
                     noteheadColor = this.rules.ColoringSetCurrent.getValue(fundamentalNote);
                 }
             }
-            if (!note.sourceNote.PrintObject) {
-                noteheadColor = transparentColor; // transparent
+            if (!note.sourceNote.PrintObject || (note.sourceNote.Notehead?.Shape === NoteHeadShape.NONE)) {
+                noteheadColor = transparentColor; // transparent (for PrintObject=false or notehead="none")
             } else if (!noteheadColor // revert transparency after PrintObject was set to false, then true again
                 || noteheadColor === "#000000" // questionable, because you might want to set specific notes to black,
                                                // but unfortunately some programs export everything explicitly as black
@@ -104,14 +105,15 @@ export class VexFlowVoiceEntry extends GraphicalVoiceEntry {
                     ", in measure #" + measureNumber);
             }*/
 
-            if (!sourceNoteNoteheadColor && this.rules.ColoringMode === ColoringModes.XML && note.sourceNote.PrintObject) {
+            if (!sourceNoteNoteheadColor && this.rules.ColoringMode === ColoringModes.XML &&
+                note.sourceNote.PrintObject && note.sourceNote.Notehead?.Shape !== NoteHeadShape.NONE) {
                 if (!note.sourceNote.isRest() && defaultColorNotehead) {
                     noteheadColor = defaultColorNotehead;
                 } else if (note.sourceNote.isRest() && defaultColorRest) {
                     noteheadColor = defaultColorRest;
                 }
             }
-            if (noteheadColor && note.sourceNote.PrintObject) {
+            if (noteheadColor && note.sourceNote.PrintObject && note.sourceNote.Notehead?.Shape !== NoteHeadShape.NONE) {
                 note.sourceNote.NoteheadColorCurrentlyRendered = noteheadColor;
             } else if (!noteheadColor) {
                 continue;
@@ -175,7 +177,7 @@ export class VexFlowVoiceEntry extends GraphicalVoiceEntry {
         }
         let stemTransparent: boolean = true;
         for (const note of this.parentVoiceEntry.Notes) {
-            if (note.PrintObject) {
+            if (note.PrintObject && note.Notehead?.Shape !== NoteHeadShape.NONE) {
                 stemTransparent = false;
                 break;
             }